feat: trail detail complete

+++ src/main/java/growthook/org/bamgang/trail/repository/TrailRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrailRepository extends JpaRepository<Trail, Long> {
+public interface TrailRepository extends JpaRepository<Trail, Integer> {
 
 }
